+++ src/pages/destinations/index.tsx
@@ -1,8 +1,8 @@
 
 import React from 'react';
-import { Helmet } from 'react-helmet';
-import { CanonicalTags } from '@/components/CanonicalTags';
 import { DestinationsHero } from '@/components/destinations/DestinationsHero';
+import { DestinationsList } from '@/components/destinations/DestinationsList';
+import { destinations } from '@/data/destinations-list';
 import { DestinationsList } from '@/components/destinations/DestinationsList';
 import { destinations } from '@/data/destinations-list';
 
@@ -17,67 +17,6 @@ const DestinationsIndex = () => {
   
   return (
     <div className="min-h-screen bg-gray-50">
-      <CanonicalTags canonicalUrl={canonicalUrl} />
-      
-      <Helmet>
-        <title>{pageTitle}</title>
-        <meta name="description" content={pageDescription} />
-        <meta name="keywords" content={pageKeywords} />
-        <meta name="robots" content="index, follow" />
-        
-        {/* Open Graph Meta Tags */}
-        <meta property="og:title" content={pageTitle} />
-        <meta property="og:description" content={pageDescription} />
-        <meta property="og:image" content={imageUrl} />
-        <meta property="og:url" content={canonicalUrl} />
-        <meta property="og:type" content="website" />
-        <meta property="og:image:secure_url" content={imageUrl} />
-        <meta property="og:locale" content="en_US" />
-        <meta property="og:site_name" content="Allergy-Free Travel" />
-        
-        {/* WhatsApp specific meta tags */}
-        <link itemProp="thumbnailUrl" href={imageUrl} />
-        <span itemProp="thumbnail" itemScope itemType="http://schema.org/ImageObject">
-          <link itemProp="url" href={imageUrl} />
-        </span>
-        
-        {/* Twitter Card Tags */}
-        <meta name="twitter:card" content="summary_large_image" />
-        <meta name="twitter:title" content={pageTitle} />
-        <meta name="twitter:description" content={pageDescription} />
-        <meta name="twitter:image" content={imageUrl} />
-        
-        {/* Schema.org markup */}
-        <script type="application/ld+json">
-          {JSON.stringify({
-            "@context": "https://schema.org",
-            "@type": "CollectionPage",
-            "mainEntity": {
-              "@type": "ItemList",
-              "itemListElement": destinations.map((dest, index) => ({
-                "@type": "ListItem",
-                "position": index + 1,
-                "item": {
-                  "@type": "TouristDestination",
-                  "name": dest.name,
-                  "description": dest.description,
-                  "url": `https://www.allergy-free-travel.com/destinations/${dest.id}`
-                }
-              }))
-            },
-            "name": pageTitle,
-            "description": pageDescription,
-            "url": canonicalUrl,
-            "dateModified": currentDate,
-            "image": {
-              "@type": "ImageObject",
-              "url": imageUrl,
-              "width": "1200",
-              "height": "630"
-            }
-          })}
-        </script>
-      </Helmet>
 
       <DestinationsHero />
       <DestinationsList />